+++ __mocks__/@etherspot/modular-sdk.js
@@ -1,9 +1,9 @@
 import * as EtherspotModular from '@etherspot/modular-sdk';
 import { ethers } from 'ethers';
 
-export const defaultAccountAddress = '0x7F30B1960D5556929B03a0339814fE903c55a347';
-export const otherFactoryDefaultAccountAddress = '0xe383724e3bDC4753746dEC781809f8CD82010914';
-export const otherAccountAddress = '0xAb4C67d8D7B248B2fA6B638C645466065fE8F1F1';
+export const defaultAccountAddressModular = '0x7F30B1960D5556929B03a0339814fE903c55a347';
+export const otherFactoryDefaultAccountAddressModular = '0xe383724e3bDC4753746dEC781809f8CD82010914';
+export const otherAccountAddressModular = '0xAb4C67d8D7B248B2fA6B638C645466065fE8F1F1';
 
 export class ModularSdk {
   sdkChainId;
@@ -17,10 +17,10 @@ export class ModularSdk {
   }
 
   getCounterFactualAddress() {
-    if (this.factoryWallet === Factory.ETHERSPOT) {
-      return defaultAccountAddress;
+    if (this.factoryWallet === 'etherspotModular') {
+      return defaultAccountAddressModular;
     }
-    return otherFactoryDefaultAccountAddress;
+    return otherFactoryDefaultAccountAddressModular;
   }
 
   async clearUserOpsFromBatch() {
@@ -64,7 +64,7 @@ export class ModularSdk {
     });
 
     return {
-      sender: defaultAccountAddress,
+      sender: defaultAccountAddressModular,
       nonce: this.nonce,
       initCode: '0x001',
       callData: '0x002',
@@ -102,16 +102,16 @@ export class ModularSdk {
   }
 
   async installModule(moduleType, module, initData, accountAddress) {
-    if (!accountAddress && !defaultAccountAddress) {
-      return 'No account address provided!'
+    if (!accountAddress && !defaultAccountAddressModular) {
+      throw new Error('No account address provided!')
     }
 
     if (!moduleType || !module) {
-      return 'installModule props missing'
+      throw new Error('installModule props missing')
     }
 
     if (module === '0x222') {
-      return 'module is already installed'
+      throw new Error('module is already installed')
     }
 
     return '0x123';
@@ -119,15 +119,15 @@ export class ModularSdk {
 
   async uninstallModule(moduleType, module, deinitData, accountAddress) {
     if (module === '0x222') {
-      return 'module is not installed'
+      throw new Error('module is not installed')
      }
 
-    if (!accountAddress && !defaultAccountAddress) {
-      return 'No account address provided!'
+    if (!accountAddress && !defaultAccountAddressModular) {
+      throw new Error('No account address provided!')
     }
 
     if (!moduleType || !module || !deinitData) {
-     return 'uninstallModule props missing'
+      throw new Error('uninstallModule props missing')
     }
 
     return '0x456';