+++ CHANGELOG.md
@@ -5,11 +5,10 @@
 ### Added Changes
 - `getSDK` include a param to choose to instantiate the Prime SDK instead of the Modular SDK
 - Added Etherspot Modular SDK `installModule` and `uninstallModule` to hook `useEtherspotModules`
-- `EtherspotContextProvider` updated with choice of instantiatiating the Prime SDK or Modular SDK
 - Added `isModular` to context `EtherspotContextProvider`
 
 ### Breaking Changes
-- Etherspot Modular SDK implemented to Transaction Kit as the default SDK
+- Etherspot Modular SDK implemented to Transaction Kit as the default `accountTemplate`
 - The account template in the `EtherspotTransactionKit` component, can only be 'etherspot' if `moddular` is true.
 
 ## [0.12.1] - 2024-05-22